Split up parser components into different modules to make files a little more manageable
