+++ npm-packages/meteor-vite/src/meteor/parser/Parser.ts
@@ -5,31 +5,20 @@ import {
     isFunctionExpression,
     isIdentifier,
     isMemberExpression,
-    isObjectExpression, isObjectProperty,
+    isObjectExpression,
+    isObjectProperty,
     isReturnStatement,
     isStringLiteral,
     Node,
-    NumericLiteral,
-    ObjectExpression,
     ObjectMethod,
     ObjectProperty,
-    StringLiteral,
     traverse,
-    VariableDeclarator,
 } from '@babel/types';
 import FS from 'fs/promises';
-import { inspect } from 'util';
-import { MeteorViteError } from '../../error/MeteorViteError';
-import Logger from '../../utilities/Logger';
-import {
-    KnownModuleMethodNames,
-    MeteorInstallCallExpression,
-    MeteorPackageProperty,
-    ModuleMethod,
-    ModuleMethodName,
-} from './ParserTypes';
+import { MeteorInstall } from './MeteorInstall';
+import { ModuleExportsError, ParserError } from './ParserError';
 
-interface ParseOptions {
+export interface ParseOptions {
     /**
      * Absolute file path to the package's JavaScript file.
      * This file needs to have been built by Meteor.
@@ -216,272 +205,7 @@ function parsePackageScope(node: Node) {
     return meteorV2(node) || meteorV3(node);
 }
 
-/**
- * Parser for a build Meteor package's meteorInstall() call.
- * Traverses through all the entries in the passed object to build up a map of each file exposed by the package
- * as well as tracking exports (module.export, module.exportDefault) and re-exports (module.link).
- * {@link https://github.com/JorgenVatle/meteor-vite/blob/8b0a7a5f5f95d78661793e8b4bc7f266c1081ed9/npm-packages/meteor-vite/test/__mocks/meteor-bundle/rdb_svelte-meteor-data.js#L25 example of meteorInstall() }
- */
-class MeteorInstall {
-    public readonly modules: ModuleList = {}
-    public readonly packageId: string;
-    public readonly name: string;
-
-    constructor({ packageId, name }: Pick<MeteorInstall, 'packageId' | 'name'>) {
-        this.packageId = packageId;
-        this.name = name;
-    }
-    
-    public static parse(node: Node) {
-        const atmosphereInstall = this.parseAtmosphereInstall(node);
-        const npmInstall = this.parseNpmInstall(node);
-        
-        if (atmosphereInstall) {
-            return atmosphereInstall;
-        }
-        return npmInstall;
-    }
-    
-    protected static parseInstall(node: MeteorInstallCallExpression) {
-        const [ modules, fileExtensions ] = node.arguments;
-        const node_modules = modules.properties[0];
-        if (propParser.getKey(node_modules) !== 'node_modules') {
-            return {
-                type: 'unknown',
-            }
-        }
-        const meteor = node_modules.value.properties[0];
-        if (propParser.getKey(meteor) === 'meteor') {
-            return {
-                type: 'atmosphere',
-                meteor,
-            }
-        }
-        return {
-            type: 'npm',
-            node_modules,
-        }
-    }
-    
-    protected static parseAtmosphereInstall(node: Node) {
-        if (!this.isRequireDeclaration(node)) return;
-        if (!this.isMeteorInstall(node.init)) return;
-        
-        
-        const { meteor, type } = this.parseInstall(node.init);
-        if (type !== 'atmosphere' || !meteor) {
-            return;
-        }
-        
-        const packageName = meteor.value.properties[0];
-        const packageModules = packageName.value.properties;
-        
-        const meteorPackage = new this({
-            packageId: `${propParser.getKey(meteor)}/${propParser.getKey(packageName)}`,
-            name: propParser.getKey(packageName),
-        });
-        
-        meteorPackage.traverseModules(packageModules, '');
-        
-        return meteorPackage;
-    }
-    
-    protected static parseNpmInstall(node: Node) {
-        if (!this.isMeteorInstall(node)) return;
-        const { node_modules, type } = this.parseInstall(node);
-        if (type !== 'npm' || !node_modules) {
-            return;
-        }
-        
-        const npmPackages = [];
-        for (const directory of node_modules.value.properties) {
-            if (!isObjectExpression(directory.value)) return; // Not a directory
-            const packageJson = {
-                name: '',
-                main: '',
-                version: '',
-            }
-            
-            for (const entry of directory.value.properties) {
-                if (!isObjectProperty(entry)) {
-                    Logger.warn(`Parsed an unexpected property from meteorInstall node module (${entry.type})`, entry);
-                    return;
-                }
-                if (propParser.getKey(entry) === 'package.json') {
-                    // Todo: fill in package.json
-                }
-                // Todo: Recurse through all directories
-                // todo: Track exported modules
-            }
-            
-            npmPackages.push(packageJson);
-        }
-    }
-    
-    protected static isRequireDeclaration(node: Node): node is VariableDeclarator {
-        if (node.type !== 'VariableDeclarator') return false;
-        if (node.id.type !== 'Identifier') return false;
-        if (node.id.name !== 'require') return false;
-        
-        return true;
-    }
-    
-    protected static isMeteorInstall(expression?: Node | null): expression is MeteorInstallCallExpression {
-        if (!expression) return false;
-        if (expression.type !== 'CallExpression') return false;
-        if (!is('Identifier', expression.callee, { name: 'meteorInstall' })) return false;
-        
-        return true;
-    }
-
-    public traverseModules(properties: MeteorPackageProperty[], parentPath: string) {
-        properties.forEach((property) => {
-            const path = `${parentPath}${property.key.value.toString()}`
-            const module = new PackageModule();
-
-
-            if (property.value.type === 'ObjectExpression') {
-                return this.traverseModules(property.value.properties, `${path}/`);
-            }
-
-            traverse(property.value.body, {
-                enter(node) {
-                    module.parse(node)
-                }
-            });
-
-            this.modules[path] = module.exports;
-        })
-    }
-    
-}
-
-/**
- * An individual file within a built Meteor package's meteorInstall file tree.
- * It essentially represents a single file and its associated exports within a given Meteor package.
- * E.g. `index.js` or `cookie-store.js`.
- *
- * {@link https://github.com/JorgenVatle/meteor-vite/blob/78a451fa311989d10cbb061bb929d8feb795ea2c/npm-packages/meteor-vite/test/__mocks/meteor-bundle/test_ts-modules.js#L20 `explicit-relative-path.ts`} would count as a module here.
- */
-class PackageModule {
-    public readonly exports: ModuleExportData[] = [];
-    // Todo: Accept module metadata from callee to provide more insightful error messages
-    constructor() {}
-
-    /**
-     * Helper for checking if an already validated node is of the provided method
-     */
-    protected isMethod<MethodName extends ModuleMethodName>(node: ModuleMethod.MethodMap[ModuleMethodName], method: MethodName): node is ModuleMethod.MethodMap[MethodName] {
-        return node.callee.property.name === method;
-    }
-
-    /**
-     * Check if provided node is a valid `module.[method]` call expression.
-     */
-    protected shouldParse(node: Node): node is ModuleMethod.MethodMap[ModuleMethodName] {
-        if (node.type !== 'CallExpression') return false;
-
-        const callee = node.callee;
-
-        if (callee.type !== 'MemberExpression') return false;
-        if (callee.object.type !== 'Identifier') return false;
-        if (!callee.object.name.match(/^module\d*$/)) return false;
-        if (callee.property.type !== 'Identifier') return false;
-        const calleeMethod = callee.property.name;
-
-        if (!KnownModuleMethodNames.includes(calleeMethod as ModuleMethodName)) {
-            Logger.warn(`Meteor module.${calleeMethod}(...) is not recognized by Meteor-Vite! Please open an issue to get this resolved! 🙏`)
-            return false
-        }
-
-        return true;
-    }
-
-    /**
-     * Parse everything within the current module and store detected exports.
-     * Todo: Possibly migrate parsers to their own class to save on memory usage?
-     */
-    public parse(node: Node) {
-        if (!this.shouldParse(node)) return;
-
-        if (this.isMethod(node, 'link')) {
-            return this.exports.push(...this.parseLink(node));
-        }
-
-        if (this.isMethod(node, 'export')) {
-            return this.exports.push(...this.parseExport(node));
-        }
-
-        if (this.isMethod(node, 'exportDefault')) {
-            this.exports.push(...this.parseExportDefault(node));
-            return;
-        }
-    }
-
-    /**
-     * Parse a Meteor bundle's `module.link()` call.
-     * {@link ModuleMethod.Link}
-     */
-    protected parseLink(node: ModuleMethod.WithoutArgs<'link'>) {
-        const [importPath, exports, id] = node.arguments;
-
-        if (importPath.type !== 'StringLiteral') {
-            throw new ModuleExportsError('Expected string as the first argument in module.link()!', importPath);
-        }
-
-        // Module.link('./some-path') without any arguments.
-        // Translates to `import './some-path' - so no exports to be found here. 👍
-        if (!exports) return [];
-
-        if (exports.type !== 'ObjectExpression') {
-            throw new ModuleExportsError('Expected ObjectExpression as the second argument in module.link()!', importPath);
-        }
-        if (id?.type !== 'NumericLiteral') {
-            throw new ModuleExportsError('Expected NumericLiteral as the last argument in module.link()!', importPath)
-        }
-
-        return formatExports({
-            packageName: importPath,
-            expression: exports,
-            id,
-        })
-    }
-
-    /**
-     * Parse a Meteor bundle's `module.export({ ... })` call.
-     * {@link ModuleMethod.export see type for examples}
-     */
-    protected parseExport(node: ModuleMethod.WithoutArgs<'export'>) {
-        if (node.arguments[0].type !== 'ObjectExpression'){
-            throw new ModuleExportsError('Unexpected export type!', exports)
-        }
-
-        return formatExports({
-            expression: node.arguments[0]
-        });
-    }
-
-    /**
-     * Parse a Meteor bundle's `module.exportDefault()` call.
-     * {@link ModuleMethod.exportDefault see examples in type declaration}
-     */
-    protected parseExportDefault(node: ModuleMethod.WithoutArgs<'exportDefault'>) {
-        const args = node.arguments;
-        let name = 'default';
-        if (args[0].type === 'Identifier') {
-            name = args[0].name;
-        } else if (args[0].type === 'ConditionalExpression') {
-            name = 'CONDITIONAL_EXPORT_DEFAULT';
-        } else {
-            new ModuleExportsError('Unexpected default export value!', args[0]);
-        }
-
-        // todo: test for default exports with `export default { foo: 'bar' }`
-        return [{ type: 'export-default', name, } satisfies ModuleExportData];
-    }
-}
-
-const propParser = {
+export const propParser = {
     getKey(property: ObjectMethod | ObjectProperty) {
         if (property.key.type === 'Identifier') {
             return property.key.name;
@@ -492,78 +216,6 @@ const propParser = {
         
         throw new ModuleExportsError('Unsupported property key type!', property);
     },
-}
-
-function formatExports({ expression, packageName, id }: {
-    expression: ObjectExpression,
-    packageName?: StringLiteral,
-    id?: NumericLiteral,
-}) {
-    return expression.properties.map((property) => {
-        if (property.type === "SpreadElement") throw new ModuleExportsError('Unexpected property type!', property);
-        const result: ModuleExportData = {
-            name: propParser.getKey(property),
-            type: 'export',
-            ...id && { id: id.value }
-        }
-        
-        if (packageName) {
-            result.type = 're-export';
-            result.from = packageName.value;
-        }
-        
-        if (result.type === 're-export' && property.type === 'ObjectMethod') {
-            result.type = 'global-binding';
-        }
-        
-        if (result.type === 're-export' && property.type === 'ObjectProperty') {
-            const content = property.value;
-            if (content.type !== 'StringLiteral') {
-                throw new ModuleExportsError('Received unsupported result type in re-export!', property);
-            }
-            
-            if (content.value !== result.name) {
-                result.as = content.value;
-            }
-        }
-        
-        return result;
-    })
-    
-}
-
-class ParserError extends MeteorViteError {
-    constructor(
-        public originalMessage: string,
-        public readonly metadata?: {
-            parseOptions?: ParseOptions,
-            node?: Node,
-        }) {
-        super(originalMessage);
-    }
-    
-    public async formatLog() {
-        const { parseOptions, node } = await this.metadata || {};
-        if (parseOptions?.fileContent) {
-            this.addLine([
-                `// File content for: ${this.metadata?.parseOptions?.filePath}`,
-                '',
-                ...(await parseOptions.fileContent).split(/[\r\n]+/)
-            ]);
-        }
-        if (node) {
-            this.addLine([inspect(node)]);
-        }
-    }
-}
-
-class ModuleExportsError extends ParserError {
-    constructor(
-        public readonly message: string,
-        public readonly node: Node
-    ) {
-        super(message, { node });
-    }
 }
 
 /**
